requirements.txt
alembic==1.8.1
asyncpg==0.26.0
attrs==22.1.0
CacheControl==0.12.11
cachy==0.3.0
certifi==2022.6.15
charset-normalizer==2.1.1
cleo==0.8.1
clikit==0.6.2
crashtest==0.3.1
distlib==0.3.6
filelock==3.8.0
greenlet==1.1.3
html5lib==1.1
idna==3.3
iniconfig==1.1.1
keyring==21.8.0
lockfile==0.12.2
Mako==1.2.2
MarkupSafe==2.1.1
msgpack==1.0.4
packaging==21.3
pastel==0.2.1
pexpect==4.8.0
pkginfo==1.8.3
platformdirs==2.5.2
pluggy==1.0.0
psycopg2==2.9.3
ptyprocess==0.7.0
py==1.11.0
pydantic==1.10.1
pylev==1.4.0
pyparsing==3.0.9
pytest==7.1.3
pytest-asyncio==0.19.0
python-dotenv==0.20.0
requests==2.28.1
requests-toolbelt==0.9.1
shellingham==1.5.0
six==1.16.0
SQLAlchemy==1.4.40
tomli==2.0.1
tomlkit==0.11.4
typing_extensions==4.3.0
urllib3==1.26.12
virtualenv==20.16.4
webencodings==0.5.1
